requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-community
sqlalchemy
psycopg2-binary
qdrant-client
sentence-transformers
python-dotenv
sqlparse
streamlit
pyngrok
pydantic